requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
pymysql==1.1.0
python-dotenv==1.0.0
flask-wtf==1.2.1
flask-login==0.6.3
flask-migrate==4.0.5